# Abordagens For loop, Sorted foreach loop e Unsafe vectorized incluídas