--- a/CSharpBenchmark/CSharpBenchmark.csproj
+++ b/CSharpBenchmark/CSharpBenchmark.csproj
@@ -19,6 +19,7 @@
     <PackageReference Include="MessagePack" Version="2.2.85" />
     <PackageReference Include="Newtonsoft.Json" Version="13.0.1" />
     <PackageReference Include="protobuf-net" Version="3.0.101" />
+    <PackageReference Include="System.IO.Pipelines" Version="5.0.1" />
     <PackageReference Include="System.Text.Json" Version="5.0.2" />
     <PackageReference Include="Utf8Json" Version="1.3.7" />
   </ItemGroup>